apps/ann-benchmarks/requirements.txt
weaviate-client[grpc]>=3.17.1
loguru==0.5.3
grpcio==1.53.0
seaborn==0.12.2
h5py==3.8.0
pandas==2.0.0
protobuf==4.21.12
